backend/requirements.txt
flask
flask-cors
python-dotenv
websocket-client
pandas
numpy
statsmodels
redis
supabase